somic/code/requirements_pip.txt
fastprogress >=0.2.2
matplotlib >=3.2.1
numpy >=1.18.1
opencv-python >=4.1.0
pandas >=1.0.3
pillow >=7.1.1
torch >=1.2.0
torchvision >=0.5.0
pyyaml >=5.3.1
albumentations >=0.4.5
tensorboard >=2.1.0
tqdm >=4.45.0
sphinx-rtd-theme
sphinx-autodoc-typehints
recommonmark 
mlflow